+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## v1.3
 
-- Added custom thumbnail support for ROM hacks, homebrew, folders, and missing artwork using exact-name matching.
+- Added custom thumbnail support for folders and all files. Improved kernel integration.